Add and style SearchResults component JSX
Replaced the placeholder content in the SearchResults
component with placeholder images and styled it to look
like the supplied design.  Used CSS columns instead of
using Flexbox to account for different image sizes.
Placeholder images are passed down to the SearchResults
component as props, but the image urls are just
temporarily coming from a static array of objects added
to the App.js file.